StrictList

Description

Definitions of strict linked list.

Most basic operations like fmap, filter, <*> can only be implemented efficiently by producing an intermediate list in reversed order and then reversing it to the original order. These intermediate reversed functions are exposed by the API, because they very well may be useful for efficient implementations of data-structures built on top of list. E.g., the "deque" package exploits them heavily.

One useful rule of thumb would be that whenever you see that a function has a reversed counterpart, that counterpart is faster and hence if you don't care about the order or intend to reverse the list further down the line, you should give preference to that counterpart.

The typical toList and fromList conversions are provided by means of the Foldable and IsList instances.

Documentation

data List a

Strict linked list.

Constructors

Cons !a !(List a)

Nil

toListReversed :: List a -> [a]

Convert to lazy list in normal form (with all elements and spine evaluated).

reverse :: List a -> List a

Reverse the list.

take :: Int -> List a -> List a

Leave only the specified amount of elements.

takeReversed :: Int -> List a -> List a

Leave only the specified amount of elements, in reverse order.

drop :: Int -> List a -> List a

Leave only the elements after the specified amount of first elements.

filter :: (a -> Bool) -> List a -> List a

Leave only the elements satisfying the predicate.

filterReversed :: (a -> Bool) -> List a -> List a

Leave only the elements satisfying the predicate, producing a list in reversed order.

takeWhile :: (a -> Bool) -> List a -> List a

Leave only the first elements satisfying the predicate.

takeWhileReversed :: (a -> Bool) -> List a -> List a

Leave only the first elements satisfying the predicate, producing a list in reversed order.

dropWhile :: (a -> Bool) -> List a -> List a

Drop the first elements satisfying the predicate.

span :: (a -> Bool) -> List a -> (List a, List a)

An optimized version of the same predicate applied to takeWhile and dropWhile. IOW,

span predicate list = (takeWhile predicate list, dropWhile predicate list)

spanReversed :: (a -> Bool) -> List a -> (List a, List a)

Same as span, only with the first list in reverse order.

break :: (a -> Bool) -> List a -> (List a, List a)

An opposite version of span. I.e.,

break predicate = span (not . predicate)

breakReversed :: (a -> Bool) -> List a -> (List a, List a)

Same as break, only with the first list in reverse order.

takeWhileFromEnding :: (a -> Bool) -> List a -> List a

Same as (takeWhile predicate . reverse). E.g.,

>>> takeWhileFromEnding (> 2) (fromList [1,4,2,3,4,5])
fromList [5,4,3]

dropWhileFromEnding :: (a -> Bool) -> List a -> List a

Same as (dropWhile predicate . reverse). E.g.,

>>> dropWhileFromEnding (> 2) (fromList [1,4,2,3,4,5])
fromList [2,4,1]

spanFromEnding :: (a -> Bool) -> List a -> (List a, List a)

Same as (span predicate . reverse).

match :: result -> (element -> List element -> result) -> List element -> result

Pattern match on list using functions.

Allows to achieve all the same as uncons only without intermediate Maybe.

Essentially provides the same functionality as either for Either and maybe for Maybe.

uncons :: List a -> Maybe (a, List a)

Get the first element and the remainder of the list if it's not empty.

head :: List a -> Maybe a

Get the first element, if list is not empty.

last :: List a -> Maybe a

Get the last element, if list is not empty.

tail :: List a -> List a

Get all elements of the list but the first one.

init :: List a -> List a

Get all elements but the last one.

initReversed :: List a -> List a

Get all elements but the last one, producing the results in reverse order.

apZipping :: List (a -> b) -> List a -> List b

Apply the functions in the left list to elements in the right one.

apZippingReversed :: List (a -> b) -> List a -> List b

Apply the functions in the left list to elements in the right one, producing a list of results in reversed order.

Reversed intermediate functions used in instances

fromListReversed :: [a] -> List a

Construct from a lazy list in reversed order.

prependReversed :: List a -> List a -> List a

Add elements of the left list in reverse order in the beginning of the right list.

mapReversed :: (a -> b) -> List a -> List b

Map producing a list in reversed order.

apReversed :: List (a -> b) -> List a -> List b

Apply the functions in the left list to every element in the right one, producing a list of results in reversed order.

explodeReversed :: (a -> List b) -> List a -> List b

Use a function to produce a list of lists and then concat them sequentially, producing the results in reversed order.

joinReversed :: List (List a) -> List a

Join (concat) producing results in reversed order.

mapMaybeReversed :: (a -> Maybe b) -> List a -> List b

Map and filter elements producing results in reversed order.

catMaybesReversed :: List (Maybe a) -> List a

Keep only the present values, reversing the order.